Portability and warning fixes in getnetgrent.c and getcap_int.c
- In getnetgrent.c, include <limits.h> inside BSDDB block
  to ensure visibility of system limits when building with DB support.
- Fix function prototype for lookup() to use UNUSED_A
  on the ypdom argument when not using YP,
  silencing unused parameter warnings with and without YP.
- In portable/getcap_int.c, fix signed/unsigned comparison warning
  by casting sizeof(pbuf) to (int)
  in the check clen < (int)sizeof(pbuf).
- These changes ensure portability and warning-free builds
  across multiple toolchains and feature flag configurations.

+++ portable/getcap_int.c
@@ -233,7 +233,7 @@ getent(char **cap, unsigned int *len, char **db_array, FILE *fp,
 			char *dbrecord;
 
 			clen = snprintf(pbuf, sizeof(pbuf), "%s.db", *db_p);
-			if (clen >= 0 && clen < sizeof(pbuf) && usedb &&
+			if (clen >= 0 && clen < (int)sizeof(pbuf) && usedb &&
 			    (capdbp = dbopen(pbuf, O_RDONLY, 0, DB_HASH, 0))) {
 				opened++;
 				retval = cdbget(capdbp, &dbrecord, name);
